Report wall-clock listened time instead of position delta (#481)

+++ app/src/main/kotlin/org/grakovne/lissen/channel/audiobookshelf/common/model/playback/ProgressSyncRequest.kt
@@ -6,6 +6,6 @@ import com.squareup.moshi.JsonClass
 @Keep
 @JsonClass(generateAdapter = true)
 data class ProgressSyncRequest(
-  val timeListened: Int,
+  val timeListened: Double,
   val currentTime: Double,
 )
